correçõe dos bugs da aplicação, implementando lógica de exeptions, reformulando todo html e css

+++ src/main/java/com/RODRIGO/RPX/services/CategoriaService.java
@@ -2,6 +2,7 @@ package com.RODRIGO.RPX.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.RODRIGO.RPX.entity.Categoria;
 import com.RODRIGO.RPX.repository.CategoriaRepository;
@@ -14,7 +15,15 @@ public class CategoriaService {
     {
         return categoriaRepository.findById(id).orElseThrow(() -> new RuntimeException("Produto não encontrado"));
     }
+    @Transactional
     public void deletar(Long id){
+         Categoria categoria = categoriaRepository.findById(id)
+            .orElseThrow(() -> new RuntimeException("Categoria não encontrada"));
+
+        if (!categoria.getProdutos().isEmpty()) {
+            throw new RuntimeException("Não é possível deletar: existem produtos associados a esta categoria.");
+        }
+
         categoriaRepository.deleteById(id);
     }
